# skeleton: delete '.input' and '.keys' files before overwriting them